Initial port of printing
Essentially we clone the original EMFormat to EMFormatHTMLPrint, add
a special PURI part that, when write_func()'ed will create HTML code
with <iframe> for each real part.

Because WebKit is somewhat buggy, this approach does not seem to be
the best since WebKit sometimes refuses to print content of <iframes>
and when frame flattening is enabled it does not print at all.

I think that the best way would be to create a single HTML file for the
entire email (like we did in Evo GtkHTML) since there are no GtkWidgets
to be printed.