Add e_web_view_{set,get}_enable_frame_flattening
By default the value is FALSE but we set it to TRUE in e_web_view_init() to force all
frames to expand to height of their content automatically.

WebKit seems to have trouble priting documents with this option enabled though, so
this function makes it easier to disable it when EWebView is used for printing.