em_utils_message_to_html() - set all validity flags found in entire processed message
Originally the EMFormat held field with flags of all detected validities
(signature/encryption). In the new EMFormat each PURI has it's own validity
flags describing validity of the particular PURI. To make the function work
as before we need to iterate through all the PURIs and return all flags we
run into.