Reimplement image loading policy (bugs #667536 and #667537)
Every http(s) request is now redirected to evo-http(s):// protocol, which we
handle manually in EMailRequest.

The EMailRequest will first try to look up the requested resource in Evo cache.
When the resource is not in the cache and image loading policy allows it to
then the resource is fetched from network. Finally, when the item is not in cache
and policy prevents EMailRequest from fetching data from intewebs then
GTK-STOCK-MISSING-IMAGE is sent.