Convert EMFormatPURI to a thread-safe EMPart object
EMFormatPURI structures are now represented by a EMPart objects which
wrap all the PURI's properties as private and make them accessible only
via em_part_{get,set}_* methods which take care of locking. Also the
EMPart is a subclass of GObject, thus it's possible to use reference
counting to ensure life-span of the object.

Unfortunately it's now a little bit more complicated to create subclasses,
because it is necessary to define a whole new class.

TODO:
 - convert image-inline, itip-formatter, mail-to-task, tnef-attachments
 	and vcard-inline plugins
 - move EMPart's private mutex to EMPartClass so that subclasses don't have
 	to have their own mutex.

Note: won't compile due to EMailDisplay not being ported