When an <iframe> is hidden, _hide() all GtkWidget it contains
When an attachment button is collapsed and a related <iframe> which contains the attachment
(message/rfc822 or text/html) is hidden, the all GtkWidget within the <iframe> must
be hidden too. Otherwise they are rendered on the WebView randomly because WebKit is unable
to anchor them to the <object> element.

<object> elements with a document (nor the document itself actually) within an <iframe> are
not considered to be hidden when the iframe itself is hidden, therefor we must use this
quite complex bindings.