Move the contact preview HTML formatter to it's own class
The EABContactFormatter provides simple public API for both synchronous
and asynchronous formatting of an EContact to HTML.

The EABContactDisplay hasn't changed from user's point of view, but
internally it only call the formatter and then just displayes the
output.

This split allows the vcard-inline plugin to render the contact compact
preview in <iframe> instead of embedding the EABContactDisplay into the
message preview pane.