Write correctly attachments with sub-PURIs
When a part is being parsed as an attachment, the parser always tries
to parse the content of the attachment, so that it can be displayed.
Sometimes, this "second parsing" can create a new PURI, effectively
a "subPURI" of the main attachment PURI.

This commit makes sure that even those "subPURIs" are written as part
of the attachment.