vcard-inline plugin: render preview as HTML in <iframe> rather then WebKitWebView widget
The vcard-inline plugin used to render the vcard preview as the
EABContactDisplay and embed it into the EMailDisplay. Since
EABContactDisplay is actually a WebKitWebView, it didn't
work much well.

This commit changes the vcard-inline plugin so that it renders
as an <iframe> and renders the preview within the <iframe> using
the new EABContactFormatter.