Make bold, italic, underline and strike-through buttons work
This is done by binding their 'activate' property with
EEditorSelection's properties for each style, so that
when selection changes, the buttons are notified about
style property change and their activate state is toggled
if neccessary.

This does not fully work for Monospaced button, as removing
of the formatting must be implemented manually.