Port spell-checking
We use our own WebKitSpellChecker, which is in general the
same as the WebKit's default, but it does not work with
language codes, but directly with EnchantDicts. This allows
us to remove ESpellDictionary and call enchant_ functions
directly from EEditor* classes.

FIXME: The spellchecking probably needs some polishing and lots
of testing...