Introduce e_editor_widget_exec_command() and port everything to it
e_editor_widget_exec_command() is a wrapper for webkit_dom_document_exec_command(),
however the command to be executed is passed as an enum member. This makes
it much easier to use, because the commands are not documented anywhere
else and the only place where to get list of all supported commands is
just from a table somewhere within WebKit source code.