Bug #689782 - [webkit-composer] Cannot start new numbered list below other one
When inserting new ordered list under existing one it was continuing in previous
one. This behavior was caused by changing the format of list but again to ordered
list (numbered, alphabetical, roman numeral lists are treated as one in WebKit).
So we have to insert these lists manually in composer through DOM modifications.