Bug #689780 - [webkit-composer] Color combo out of sync with text color
Actually all the composer settings regarding to fonts (color, boldness, ...)
were not re-set when the format was changed. Also detection of some font properities
as e.g. boldness was rewritten for usage of CSS properities instead of looking for
element tags for these properities (e.g. <b> for boldness). The old system was not
working properly in some cases e.g. when we are setting the font format as bold
and after that to italic it could end with <b> tag with style attribute set with
font-style=italic entry. So in that case the lookup for <i> tag will fail. Also some
code style changes were done.