EEditorSelection: Fix unindentation
When we have this text with indentation in composer:
first_line
   a

   a

   X

And we have cursor instead of X and we will try to unindent that line
it will remove all after the first a. We have to mark the element that
we want to unindent (with -x-evo-to-unindent class) to prevent
webkit_node_is_equal_node return false positives when it will hit other empty
lines like in this case.