When generating the HTML code of the message write the additional data after the message
Before this change we were putting BR elements and also the credits
before the actual HTML code of the message. But this was wrong as when
WebKit was loading the given HTML code that looked like
<br>CREDITS<html>MESSAGE_CODE</html> WebKit parsed it like
<html><br>CREDITS</html><html>MESSAGE_CODE</html>. As no elements are
allowed outside of HTML root element WebKit wrapped them into another
HTML root element. Afterwards the first root element was treated as the
primary one and all the elements from the second's root HEAD and BODY
elements were moved from to the first one. Thus the HTML that was loaded
into composer contained the i.e. META or STYLE definitions in the body.
With this change we are putting and wrapping the elements behind the
message and afterwards moving them into the body. Also when we need to
cite the whole body of the message we're appending the span with
-x-evo-cite-body class to the message.