Hide insensitive paragraph styles in the Plain Text mode
Instead of showing a complete list of paragraph styles with some
insensitive in the Plain Text mode, hide those insensitive completely.
It makes the UI nicer and doesn't confuse users with options, which
are not applicable in the Plain Text mode, the same way as the HTML
toolbar is hidden.